% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_currency.R
\name{convert_currency}
\alias{convert_currency}
\title{This function prints forint formatted.}
\usage{
convert_currency(incurrency, outcurrency, days)
}
\arguments{
\item{incurrency}{is the currency you want to convert from}

\item{outcurrency}{is the currency you want to convert to}

\item{days}{is the amount of days you want to see the conversion rate}
}
\description{
This function prints forint formatted.
}
