% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{This gets the current bitcoin price in USD from Binance}
\usage{
get_bitcoin_price(retried = 0)
}
\description{
This gets the current bitcoin price in USD from Binance
}
