% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bitcoin_price.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Gets the most recent bitcoin price in USD from Binance API}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{the number of retries previously done before the exponential backoff sleep}
}
\description{
This is a wrapper around a \code{library(daroczig/binancer)} function \code{binance_coins_prices}
}
